Secure Stripe Connect secrets
