+++ apps/web/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@hello-pangea/dnd": "^18.0.1",
     "@heroicons/react": "^2.1.5",
+    "@google-cloud/secret-manager": "^5.6.0",
     "@react-pdf/renderer": "^3.4.4",
     "@stripe/react-stripe-js": "^2.7.1",
     "@stripe/stripe-js": "^4.10.0",
