+++ functions/package.json
@@ -14,6 +14,7 @@
     "@ffmpeg-installer/ffmpeg": "^1.1.0",
     "color-thief-node": "^1.0.4",
     "cors": "^2.8.5",
+    "@google-cloud/secret-manager": "^5.6.0",
     "firebase-admin": "^12.5.0",
     "firebase-functions": "^5.0.1",
     "fluent-ffmpeg": "^2.1.2",
